pyproject.toml
[tool.poetry]
name = "lacrei"
version = "0.1.0"
description = "Projeto desafio backend para a Lacrei Saude"
authors = ["jon1nline <www.jhonattan@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "lacreisaude" }]


[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.2.2,<6.0.0"
drf-yasg = ">=1.21.10,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
markdown = ">=3.8,<4.0"
django-filter = ">=25.1,<26.0"
psycopg2 = "^2.9.10"
pyjwt = "^2.10.1"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"